--- a/src/shortestpaths/bellman-ford.jl
+++ b/src/shortestpaths/bellman-ford.jl
@@ -37,31 +37,35 @@ function bellman_ford_shortest_paths(
     distmx::AbstractMatrix{T}=weights(graph)
     ) where T<:Real where U<:Integer
 
-    active = Set{U}(sources)
-    sizehint!(active, nv(graph))
+    active = falses(nv(graph))
+    for s in sources
+        active[s] = true
+    end
     dists = fill(typemax(T), nv(graph))
     parents = zeros(U, nv(graph))
     dists[sources] .= 0
     no_changes = false
+    new_active = falses(nv(graph))
 
     for i in one(U):nv(graph)
         no_changes = true
-        new_active = Set{U}()
-        for u in active
+        new_active.chunks .= 0
+        active_vertices = U[v for v in findall(active)]
+        for u in active_vertices
             for v in outneighbors(graph, u)
                 relax_dist = distmx[u, v] + dists[u]
                 if dists[v] > relax_dist
                     dists[v] = relax_dist
                     parents[v] = u
                     no_changes = false
-                    push!(new_active, v)
+                    new_active[v] = true
                 end
             end
         end
         if no_changes
             break
         end
-        active = new_active
+        active.chunks .= new_active.chunks
     end
     no_changes || throw(NegativeCycleError())
     return BellmanFordState(parents, dists)