Almost a 2X speedup and some allocations are saved.

**BENCHMARKS** 

NEW -
```julia
julia> g1 = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> g2 = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> g3 = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> g4 = loadsnap(:email_enron_full)
{36692, 183831} undirected simple Int64 graph

julia> @benchmark bellman_ford_shortest_paths(g1, x) setup=(x=rand(1:nv(g1)))
BenchmarkTools.Trial: 
  memory estimate:  1.26 MiB
  allocs estimate:  14
  --------------
  minimum time:     247.723 μs (0.00% GC)
  median time:      14.191 ms (0.00% GC)
  mean time:        11.883 ms (0.43% GC)
  maximum time:     15.261 ms (0.00% GC)
  --------------
  samples:          421
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g2, x) setup=(x=rand(1:nv(g2)))
BenchmarkTools.Trial: 
  memory estimate:  129.06 KiB
  allocs estimate:  22
  --------------
  minimum time:     605.471 μs (0.00% GC)
  median time:      686.534 μs (0.00% GC)
  mean time:        694.139 μs (0.76% GC)
  maximum time:     1.883 ms (47.91% GC)
  --------------
  samples:          7114
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g3, x) setup=(x=rand(1:nv(g3)))
BenchmarkTools.Trial: 
  memory estimate:  2.53 MiB
  allocs estimate:  34
  --------------
  minimum time:     10.741 ms (0.00% GC)
  median time:      12.215 ms (0.00% GC)
  mean time:        12.221 ms (0.49% GC)
  maximum time:     13.204 ms (4.03% GC)
  --------------
  samples:          409
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g4, x) setup=(x=rand(1:nv(g4)))
BenchmarkTools.Trial: 
  memory estimate:  583.38 KiB
  allocs estimate:  16
  --------------
  minimum time:     158.162 μs (0.00% GC)
  median time:      4.160 ms (0.00% GC)
  mean time:        3.897 ms (0.71% GC)
  maximum time:     5.309 ms (0.00% GC)
  --------------
  samples:          1279
  evals/sample:     1

```

OLD -
```julia
julia> @benchmark bellman_ford_shortest_paths(g1, x) setup=(x=rand(1:nv(g1)))
BenchmarkTools.Trial: 
  memory estimate:  2.37 MiB
  allocs estimate:  23
  --------------
  minimum time:     263.942 μs (0.00% GC)
  median time:      31.321 ms (0.00% GC)
  mean time:        28.654 ms (0.34% GC)
  maximum time:     33.966 ms (1.43% GC)
  --------------
  samples:          175
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g2, x) setup=(x=rand(1:nv(g2)))
BenchmarkTools.Trial: 
  memory estimate:  201.92 KiB
  allocs estimate:  87
  --------------
  minimum time:     1.129 ms (0.00% GC)
  median time:      1.202 ms (0.00% GC)
  mean time:        1.217 ms (0.81% GC)
  maximum time:     2.363 ms (38.46% GC)
  --------------
  samples:          4074
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g3, x) setup=(x=rand(1:nv(g3)))
BenchmarkTools.Trial: 
  memory estimate:  4.14 MiB
  allocs estimate:  152
  --------------
  minimum time:     24.306 ms (0.00% GC)
  median time:      28.425 ms (0.00% GC)
  mean time:        28.021 ms (0.47% GC)
  maximum time:     30.961 ms (0.00% GC)
  --------------
  samples:          179
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g4, x) setup=(x=rand(1:nv(g4)))
BenchmarkTools.Trial: 
  memory estimate:  1.12 MiB
  allocs estimate:  28
  --------------
  minimum time:     127.651 μs (0.00% GC)
  median time:      9.186 ms (0.00% GC)
  mean time:        8.812 ms (0.70% GC)
  maximum time:     19.234 ms (0.00% GC)
  --------------
  samples:          567
  evals/sample:     1

```