@abhinavmehndiratta 2019-04-07
Okay for some reason I thought the earlier approach would be faster (I was wrong). Here are the updated benchmarks -
```julia
julia> g1 = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> g2 = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> g3 = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> g4 = loadsnap(:email_enron_full)
{36692, 183831} undirected simple Int64 graph

julia> @benchmark bellman_ford_shortest_paths(g1, x) setup=(x=rand(1:nv(g1)))
BenchmarkTools.Trial: 
  memory estimate:  1.26 MiB
  allocs estimate:  13
  --------------
  minimum time:     246.867 μs (0.00% GC)
  median time:      14.570 ms (0.00% GC)
  mean time:        12.936 ms (0.30% GC)
  maximum time:     25.391 ms (0.00% GC)
  --------------
  samples:          386
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g2, x) setup=(x=rand(1:nv(g2)))
BenchmarkTools.Trial: 
  memory estimate:  96.88 KiB
  allocs estimate:  17
  --------------
  minimum time:     593.677 μs (0.00% GC)
  median time:      648.031 μs (0.00% GC)
  mean time:        659.896 μs (0.56% GC)
  maximum time:     1.822 ms (48.24% GC)
  --------------
  samples:          7477
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g3, x) setup=(x=rand(1:nv(g3)))
BenchmarkTools.Trial: 
  memory estimate:  1.90 MiB
  allocs estimate:  23
  --------------
  minimum time:     10.591 ms (0.00% GC)
  median time:      12.590 ms (0.00% GC)
  mean time:        12.973 ms (0.36% GC)
  maximum time:     24.223 ms (2.82% GC)
  --------------
  samples:          385
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g4, x) setup=(x=rand(1:nv(g4)))
BenchmarkTools.Trial: 
  memory estimate:  583.19 KiB
  allocs estimate:  14
  --------------
  minimum time:     157.811 μs (0.00% GC)
  median time:      3.921 ms (0.00% GC)
  mean time:        3.635 ms (0.59% GC)
  maximum time:     5.214 ms (0.00% GC)
  --------------
  samples:          1370
  evals/sample:     1

```