@sbromberger 2019-04-07
This looks great. Could you please post some final benchmarks?