@abhinavmehndiratta 2019-04-07
Swapping does help a bit.
```julia
julia> g1 = loadsnap(:ego_twitter_d)
{81306, 1768149} directed simple Int64 graph

julia> g2 = loadsnap(:facebook_combined)
{4039, 88234} undirected simple Int64 graph

julia> g3 = loadsnap(:soc_slashdot0902_u)
{82168, 582533} undirected simple Int64 graph

julia> g4 = loadsnap(:email_enron_full)
{36692, 183831} undirected simple Int64 graph

julia> @benchmark bellman_ford_shortest_paths(g1, x) setup=(x=rand(1:nv(g1)))
BenchmarkTools.Trial: 
  memory estimate:  1.26 MiB
  allocs estimate:  18
  --------------
  minimum time:     228.525 μs (0.00% GC)
  median time:      13.825 ms (0.00% GC)
  mean time:        12.219 ms (0.38% GC)
  maximum time:     19.336 ms (0.00% GC)
  --------------
  samples:          409
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g2, x) setup=(x=rand(1:nv(g2)))
BenchmarkTools.Trial: 
  memory estimate:  97.23 KiB
  allocs estimate:  30
  --------------
  minimum time:     581.846 μs (0.00% GC)
  median time:      667.465 μs (0.00% GC)
  mean time:        708.236 μs (0.62% GC)
  maximum time:     2.402 ms (64.27% GC)
  --------------
  samples:          6955
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g3, x) setup=(x=rand(1:nv(g3)))
BenchmarkTools.Trial: 
  memory estimate:  1.90 MiB
  allocs estimate:  42
  --------------
  minimum time:     10.164 ms (0.00% GC)
  median time:      11.923 ms (0.00% GC)
  mean time:        12.493 ms (0.38% GC)
  maximum time:     24.867 ms (0.00% GC)
  --------------
  samples:          400
  evals/sample:     1

julia> @benchmark bellman_ford_shortest_paths(g3, x) setup=(x=rand(1:nv(g3)))

BenchmarkTools.Trial: 
  memory estimate:  1.90 MiB
  allocs estimate:  42
  --------------
  minimum time:     10.089 ms (0.00% GC)
  median time:      12.030 ms (0.00% GC)
  mean time:        12.340 ms (0.38% GC)
  maximum time:     22.063 ms (2.73% GC)
  --------------
  samples:          405
  evals/sample:     1


julia> @benchmark bellman_ford_shortest_paths(g4, x) setup=(x=rand(1:nv(g4)))
BenchmarkTools.Trial: 
  memory estimate:  583.36 KiB
  allocs estimate:  21
  --------------
  minimum time:     95.215 μs (0.00% GC)
  median time:      3.873 ms (0.00% GC)
  mean time:        3.609 ms (0.62% GC)
  maximum time:     5.951 ms (10.27% GC)
  --------------
  samples:          1381
  evals/sample:     1

```